refactor : ConnectionManager 삭제
